cluster: fix bug in Cluster.add_nodes
Needed to also check if Cluster.spot_bid was specified when checking
whether to wait on instances or spot requests to propagate. The spot_bid
kwarg to add_nodes only enables spot if it's not already or overrides
the existing spot bid - it can not disable the global Cluster.spot_bid.

closes gh-284