ebsimage: log the ebs volume formatting step
This prevents users from thinking the ebsimage command has stalled when
using it on an instance-store instance. In this case the ebsimage
command has to attach and format a volume and then sync the root file
system which is a lengthy process. The lack of info message for the
formatting step leaves users in the dark while the format takes place.
Added an info message and also set silent=False when running mkfs.ext3
and rsync so that users are more aware of what's happening.

closes gh-334