spothistory: use classic or vpc prices based on default vpc
Updated EasyEC2.get_spot_history to either show VPC or EC2-classic spot
prices based on whether or not the user has a default vpc. If the user
does have a default VPC then the default spot prices are VPC otherwise
it defaults to EC2-classic prices. Also added vpc=False and
classic=False kwargs to EasyEC2.get_spot_history which forces the
function to return prices for a given platform regardless of whether or
not the user has a default VPC.