node: ignore exit status when starting portmap
This was needed for old AMIs in order for NFS to start-up properly.
Updated the code to ignore any error starting portmap when running this
command given that not all AMIs have/need it which avoids a fatal error
in these cases. The important service in this code is NFS which will
raise a fatal error if it fails.