awsutils: raise exc in wait_for_propagation after max_retries
EasyEC2.wait_for_proagation now raises an exception if the objects are
not able to be refetched after max_retries.

Increased max_retries to 60 in order to give enough time (5mins by
default which should be overkill) for the API to return the desired
objects (ie spot or instance requests) before raising the exception. For
some reason the filters API seems to take longer than the query list
method to return spot requests in some regions (e.g. us-west-2). For
example, from limited testing at the time of this commit it seems the
following command in us-west-2 region:

$ aws ec2 describe-spot-instance-requests --spot-instance-request-ids \
sir-blahblah2

takes longer to return the newly created spot request than the equivalent
'filter' version of the command:

$ aws ec2 describe-spot-instance-requests --filters \
Name=spot-instance-request-id,Values=sir-blahbla2