spothistory: use history data for plot's date range heading
This more accurately shows the date range that was returned by Amazon
for the start and end times requested by the user. It also ensures the
plot shows the date range regardless of whether the user specified an
exact range, an approximate range (e.g. --days-ago) or no range at all.