spothistory: improve date calcs for -s, -e, and -d options
Updated the code to pass start=None and end=None if no specific ranges
were specified (ie -s, -e, and -d options). When -d is specified only
the start time is calculated using local time (converted to UTC) -
end=None is still used unless the user explicitly sets -e option in
which case the start time is calculated based on the end time minus the
number of days specified by -d. This avoids discrepancies between the
AWS console starcluster in the case that the user's local time is off.

Added a check to error out if both -d and -s options are specified.

Added a note to the spothistory web template that all timestamps
displayed in the plot are UTC.