removenode: update usage to be similar to addnode
The removenode command can now be used similar to the addnode command,
e.g.:

$ starcluster rn mycluster # selects and removes a single node
$ starcluster rn mycluster -n 3 # selects and removes three nodes
$ starcluster rn mycluster -a node001,node002 # remove specific nodes

Deprecated the old usage where users passed a list of node aliases to
remove as arguments, e.g.:

$ starcluster rn mycluster node001 node002 node003

This still works but prompts a warning for now. Support for this
deprecated usage will be removed in a future release.