setup: workaround for pytest plugin loading bug
For some reason the pytest_cache plugin is not being loaded properly
when using 'python setup.py test' to manage the test dependencies. If
all dependencies are installed in the site-packages directory this issue
doesn't occur. This issue seems to occur when test deps are missing and
therefore are fetched locally as eggs by 'python setup.py test'. The fix
is to manually call configure_setuptools_entrypoints() on pytest's
PluginManager object in order to properly register the plugin(s). An
alternate fix is to import pytest_cache and pass it to pytest's main
function's plugins kwarg.