start: set default value to None for --no-public-ips
By default StarCluster's start command interprets any options that have
a non-None value to be user-specified options. If the default value of
an option is not None then the start command assumes the default value
to be a user specified value and overrides the config with command-line
option values.

In this case the --no-public-ip default value was not None so setting
PUBLIC_IPS in the config had no effect (unless it happened matched the
start command's default for it of course :P) due to the start command
overriding it with the non-None default value.

closes gh-366