cluster: improve wait_for_active_spots
Require active spots to also have an instance_id allocated before
exiting the wait loop. This ensures that the spot requests are 'active'
and also have an instance ID that we can use in the
'wait_for_propagation' call later on. This *might* fix
InstanceDoesNotExist errors related to eventual consistency when using
spot nodes (gh-377).