cluster: fix filters in spot_requests property
Updated the VPC check in Cluster.spot_requests to make sure the nodes
were explicitly launched in a VPC by StarCluster. This is needed in
order to make sure the spot request filters are correct in the
Cluster.spot_requests property. This is because StarCluster launches VPC
nodes differently than non-VPC (in order to facilitate the public_ips
                option) which changes the filters needed in order to
correctly account for the instances. The previous check assumed that if
the cluster group had a vpc_id attribute then StarCluster must have
launched it, however, this is not true for the default VPC. In this case
StarCluster launches the clusters with a non-VPC configuration but the
default VPC automagically adds the group and instances to the default
VPC and subnet.

closes gh-377