sge: run qconf commands on exec host in _inst_sge
This avoids a threadpool issue where the same SSH connection object
can't be called by more than one thread. Each node in the cluster is an
admin host by default so the same qconf commands in _inst_sge can be
directly executed on the node being added as an exec host which avoids
the threadpool issue.