tests: fix for latest changes to pytest_cov
It seems with the latest changes to pytest_cov it's now impossible to
configure the coverage plugin exclusively from conftest.py without
manually loading and configuring the plugin in pytest_configure. Without
this the coverage plugin can ONLY be loaded/configured via the command
line arguments to pytest (which sucks btw). This change to pytest_cov
broke the build on travis-ci - this hack should fix that.