Make loadbalancer account for master instance type
The load balancer currently assumes that all nodes in a cluster
(including the master node) are of the same instance type and throws an
exception if this is not so. But the StarCluster config explicitly
allows the master node to be of a different instance type than its
execution hosts. Therefore if a cluster is configured in such a manner
the loadbalancer will fail to run.

This commit makes the loadbalancer take into account that the master
node can be of a different instance type.